Moved LoggingContextFilter into a module
