+++ src/loggingex/context/__init__.py
@@ -5,16 +5,16 @@ LoggingContextFilter is class that you need to add to your loggers or handlers.
 The context helper is function that you can use anywhere in your code to quickly
 put value into the logging context.
 """
-from logging import LogRecord
-
-from .change import CONTEXT_STORE_CLASS, ContextChange
+from .change import ContextChange
 from .exceptions import (
     ContextChangeAlreadyStartedException,
     ContextChangeNotStartedException,
     ContextException,
     ContextInvalidNameException,
 )
-from .store import CONTEXT_STORE_VARIABLE_NAME, ContextStore
+from .filter import LoggingContextFilter
+from .store import ContextStore
+
 
 __all__ = (
     # exceptions
@@ -25,53 +25,10 @@ __all__ = (
     # internal-ish classes
     "ContextStore",
     "ContextChange",
-    # settings
-    "CONTEXT_STORE_VARIABLE_NAME",
-    "CONTEXT_STORE_CLASS",
     # public api
     "LoggingContextFilter",
     "context",
 )
-
-
-class LoggingContextFilter:
-    """Logging filter injects current context variables into the log records."""
-
-    IGNORED_VARIABLE_NAMES = (
-        "name",
-        "msg",
-        "args",
-        "levelname",
-        "levelno",
-        "pathname",
-        "filename",
-        "module",
-        "exc_info",
-        "exc_text",
-        "stack_info",
-        "lineno",
-        "funcName",
-        "created",
-        "msecs",
-        "relativeCreated",
-        "thread",
-        "threadName",
-        "processName",
-        "process",
-    )
-
-    def filter(self, record: LogRecord):  # noqa: A003
-        """Inject current context variables into the record.
-
-        :param record: LogRecord to inject context into.
-        :return: Always returns 1.
-        """
-        context_store = CONTEXT_STORE_CLASS()
-        context = context_store.get()
-        for name, value in context.items():
-            if name not in self.IGNORED_VARIABLE_NAMES:
-                setattr(record, name, value)
-        return 1
 
 
 class _ContextChangeShortcuts:
